Fix: firebase 메소드를 컴포넌트와 분리
