+++ src/components/SendMessage/SendMessage.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { db } from '../../firebase/config';
-import { addDoc, collection, serverTimestamp } from 'firebase/firestore';
+import { createMessages } from '../../firebase/createMessages';
 import shortId from 'shortid';
 import getNickname from 'api/getNickname';
 import * as styled from './SendMessage.styles';
@@ -18,29 +17,17 @@ async function saveUserInfoToLocalStorage() {
   localStorage.setItem('uid', uid);
   localStorage.setItem('displayName', displayName);
 }
-function getUserInfoFromLocalStorage() {
-  const uid = localStorage.getItem('uid');
-  const displayName = localStorage.getItem('displayName');
-  return { uid, displayName };
-}
 
 const SendMessage = () => {
   const [message, setMessage] = useState('');
 
-  const sendMessage = async (event: any) => {
+  const handleSendMessage = async (event: any) => {
     event.preventDefault();
     if (message.trim() === '') {
       alert('메시지를 입력해주세요');
       return;
     }
-    const { uid, displayName } = getUserInfoFromLocalStorage();
-
-    await addDoc(collection(db, 'messages'), {
-      text: message,
-      name: displayName,
-      createdAt: serverTimestamp(),
-      uid,
-    });
+    createMessages(message);
     setMessage('');
   };
 
@@ -53,7 +40,7 @@ const SendMessage = () => {
   }, []);
 
   return (
-    <styled.SendMessageForm onSubmit={(event) => sendMessage(event)}>
+    <styled.SendMessageForm onSubmit={(event) => handleSendMessage(event)}>
       <styled.SendMessageInput
         name="messageInput"
         type="text"
